move Test_truncateWeek to time_test.go
